Introduce ExtractorCreator class.
