+++ mikeio1d/result_extractor/__init__.py
@@ -1,3 +1,6 @@
+from .extractor_creator import ExtractorCreator
+from .extractor_creator import ExtractorOutputFileType
+
 from .extractor import Extractor
 from .extractor_all import ExtractorAll
 from .extractor_csv import ExtractorCsv
