+++ mikeio1d/result_extractor/extractor_all.py
@@ -3,14 +3,7 @@ from .extractor_csv import ExtractorCsv
 from .extractor_txt import ExtractorTxt
 
 
-class OutputFileType(object):
-    TXT = 'txt'
-    CSV = 'csv'
-    DFS0 = 'dfs0'
-    ALL = '-'
-
-
-class ExtractorAll(object):
+class ExtractorAll:
     """Class which extracts data into all supported file formats"""
 
     def __init__(self, out_file_name, output_data, result_data, time_step_skipping_number=1):
@@ -19,19 +12,6 @@ class ExtractorAll(object):
             ExtractorCsv(out_file_name.replace(".-", ".csv"), output_data, result_data, time_step_skipping_number),
             ExtractorDfs0(out_file_name.replace(".-", ".dfs0"), output_data, result_data, time_step_skipping_number)
         ]
-
-    @staticmethod
-    def create(out_file_type, out_file_name, output_data, result_data, time_step_skipping_number=1):
-        extractors = {
-            OutputFileType.TXT: ExtractorTxt,
-            OutputFileType.CSV: ExtractorCsv,
-            OutputFileType.DFS0: ExtractorDfs0,
-            OutputFileType.ALL: ExtractorAll
-        }
-
-        extractor = extractors.get(out_file_type, None)
-
-        return extractor(out_file_name, output_data, result_data, time_step_skipping_number)
 
     def export(self):
         for extractor in self.all_extractors:
